Yet more fixes

+++ tests/convert_utf32_to_latin1_with_errors_tests.cpp
@@ -41,7 +41,7 @@ TEST(convert_latin1_only) {
 
 TEST(convert_fails_if_input_too_large) {
   for (size_t j = 0; j < trials; j++) { 
-    uint32_t seed = std::chrono::system_clock::now().time_since_epoch().count();
+    uint32_t seed = static_cast<uint32_t>(std::chrono::system_clock::now().time_since_epoch().count());
     simdutf::tests::helpers::RandomInt generator(0xFF, 0xffffffff, seed);
     const size_t size = 64;
     simdutf::tests::helpers::transcode_utf32_to_latin1_test_base test([](){ return '*'; }, size+32);
